lib(): mv helper to separate file, make module.exports a named function
