+++ lib/index.js
@@ -1,37 +1,28 @@
-function walk (options, nodes) {
-  nodes.map((node) => {
-    if (typeof node !== 'object') return
-    if (node.attrs) {
-      const opt = intersectKeys(options, node.attrs)[0]
-      if (opt) {
-        delete node.attrs[opt]
-        node.content = [options[opt](node.content[0])]
+// --------------------------------
+// #POSTHTML - CONTENT
+// ------------------------------------
+
+'use strict'
+
+const intersectKeys = require('./intersectKeys')
+
+module.exports = function PostHTMLContent (options, cb) {
+  return function content (options, nodes) {
+    nodes.map((node) => {
+      if (typeof node !== 'object') return
+
+      if (options[node.tag]) {
+        const option = node.tag
+        node.content = [options[option](node.content[0])]
       }
-    }
-    walk(options, node.content)
-  })
+
+      if (node.attrs) {
+        const option = intersectKeys(options, node.attrs)[0]
+        delete node.attrs[option]
+        node.content = [options[option](node.content[0])]
+      }
+
+      content(options, node.content)
+    })
+  }.bind(null, options)
 }
-
-// helper: because it's very fast and doesnt need a dependency
-function intersectKeys (_a, _b) {
-  let ai = 0
-  let bi = 0
-  let a = Object.keys(_a)
-  let b = Object.keys(_b)
-  const result = []
-
-  while (ai < a.length && bi < b.length) {
-    if (a[ai] < b[bi]) {
-      ai++
-    } else if (a[ai] > b[bi]) {
-      bi++
-    } else {
-      result.push(a[ai])
-      ai++
-      bi++
-    }
-  }
-  return result
-}
-
-module.exports = (options) => walk.bind(null, options)
